src/requirements.txt
domeneshop
azure-functions
requests